+++ Packages/com.yontalane.aseprite/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.3] - 2025.07.22
+
+### Changed
+
+- MotionReceiver is now AnimationBridge
+
+### Added
+
+- AnimationBridge can be used to play animations
+- AnimationBridge can be used to detect current animation
+
 ## [1.0.2] - 2025.07.22
 
 ### Added